% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\docType{class}
\name{BaseClass-class}
\alias{BaseClass-class}
\title{BaseClass}
\description{
A BaseClass is an abstract class. Sub classes are inherited from BaseClass.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) name of the object}

\item{\code{valuevector}}{(numeric) constructed feature vector}

\item{\code{isvalid}}{(logical) result of object validation: checksubclassobjectvalidity}

\item{\code{preimpute}}{(logical) whether valuevector should be compute before missing value imputation}
}}

