% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{checksubclassobjectvalidity}
\alias{checksubclassobjectvalidity}
\title{checksubclassobjectvalidity}
\usage{
checksubclassobjectvalidity(object, dataobject)
}
\arguments{
\item{object}{A user-defined or default sub class object inhereted from BaseClass}

\item{dataobject}{An object of DataClass}
}
\value{
(numeric) 1 for valid objects

(numeric) 0 for not valid object
}
\description{
checksubclassobjectvalidity is a validation method for BaseClass objects.
Valid constructed feature vectors must be finite, have variance and be of same length as rows in data.
}

