% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{computeValue}
\alias{computeValue}
\title{computeValue}
\usage{
computeValue(object, dataobject)
}
\arguments{
\item{object}{A sub class object inheriting BaseClass}

\item{dataobject}{A DataClass object}
}
\value{
(numeric) feature vector
}
\description{
computevalue is a BaseClass generic function for computing constructed feature vectors.
}

