% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{constructfeature}
\alias{constructfeature}
\title{constructor function for adding constructed features to the system}
\usage{
constructfeature(classname, operation, mode = "all", impute = FALSE)
}
\arguments{
\item{classname}{(character) name of the inherited class}

\item{operation}{(expression) feature construction operation.The expression is evaluated by computevalue method.}

\item{mode}{(character) Mode of data to be used in construction . Defaults to "all", option "numeric" for numeric data without class labels.}

\item{impute}{(logical) Impute whether construction is done before missing value imputation . Defaults to "FALSE"}
}
\value{
(BaseClass) A sub class inhereted from BaseClass and computevalue method for the class
}
\description{
constructor function for adding constructed features to the system
}

