% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{ControlClass-class}
\alias{ControlClass-class}
\title{An S4 class representing setups to be executed}
\description{
An S4 class representing setups to be executed
}
\section{Slots}{

\describe{
\item{\code{setups}}{A list of SetUpClass objects}
}}

