% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{DataClass-class}
\alias{DataClass-class}
\title{An S4 class representing data objects.}
\description{
DataClass object can be initialized only for a data frame that has a) one class label columns of class 'factor' and
b) other columns are of type 'numeric'
}
\section{Slots}{

\describe{
\item{\code{name}}{(character) name of the setup object}

\item{\code{basedata}}{(data frame) original data to be visualized}

\item{\code{imputedbase}}{(data frame) missing value in original data imputed with Knn imputation}

\item{\code{numericdata}}{(data frame) numeric columns of original data}

\item{\code{imputednumeric}}{(data frame) imputed numeric columns}
}}

