% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DefaultControl.R
\docType{data}
\name{defaultParameters}
\alias{defaultParameters}
\title{defaultParameters}
\format{\preformatted{Formal class 'ParameterClass' [package "preproviz"] with 1 slot
  ..@ parameters:List of 10
  .. ..$ : chr "MissingValueShare"
  .. ..$ : chr "MissingValueToClass"
  .. ..$ : chr "LOFScore"
  .. ..$ : chr "MahalanobisToClassCenter"
  .. ..$ : chr "DistanceToNearest"
  .. ..$ : chr "LenghtOfIQR"
  .. ..$ : chr "NearestPointPurity"
  .. ..$ : chr "ClassificationCertainty"
  .. ..$ : chr "NeighborhoodDiversity"
  .. ..$ : chr "ScatterCounter"
}}
\usage{
defaultParameters
}
\description{
defaultParameters include nine experimental constructed features (techically, subclasses)
}
\keyword{datasets}

