% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getbasedata}
\alias{getbasedata}
\alias{getbasedata,AnalysisClass-method}
\alias{getbasedata,RunClass-method}
\title{getbasedata}
\usage{
getbasedata(object)

\S4method{getbasedata}{AnalysisClass}(object)

\S4method{getbasedata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass)}
}
\value{
(data frame) basedata
}
\description{
get basedata, that is the original data frame to be visualized
}
\keyword{internal}

