% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getclasslabels}
\alias{getclasslabels}
\alias{getclasslabels,AnalysisClass-method}
\alias{getclasslabels,RunClass-method}
\title{getclasslabels}
\usage{
getclasslabels(object)

\S4method{getclasslabels}{AnalysisClass}(object)

\S4method{getclasslabels}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(factor) vector of class labels
}
\description{
get class labels of basedata
}
\keyword{internal}

