% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getcmdsdata}
\alias{getcmdsdata}
\alias{getcmdsdata,AnalysisClass-method}
\title{get classical multidimensional scaling from minmaxconstructed data}
\usage{
getcmdsdata(object)

\S4method{getcmdsdata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) cmds points
}
\description{
get classical multidimensional scaling from minmaxconstructed data
}
\keyword{internal}

