% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getcombineddata}
\alias{getcombineddata}
\alias{getcombineddata,AnalysisClass-method}
\alias{getcombineddata,RunClass-method}
\title{get basedata and constructed data combined}
\usage{
getcombineddata(object)

\S4method{getcombineddata}{AnalysisClass}(object)

\S4method{getcombineddata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) combined data
}
\description{
get basedata and constructed data combined
}
\keyword{internal}

