% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getconstructeddata}
\alias{getconstructeddata}
\alias{getconstructeddata,AnalysisClass-method}
\alias{getconstructeddata,RunClass-method}
\title{getconstructeddata}
\usage{
getconstructeddata(object)

\S4method{getconstructeddata}{AnalysisClass}(object)

\S4method{getconstructeddata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) constructed data
}
\description{
get constructeddata, that is features constructed combined

getconstructeddata AnalysisClass

getconstructeddata RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 

\item \code{RunClass}: 
}}

