% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getdendrodata}
\alias{getdendrodata}
\alias{getdendrodata,AnalysisClass-method}
\title{getdendrodata (NOT IN USE)}
\usage{
getdendrodata(object)

\S4method{getdendrodata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) dendro data
}
\description{
get dendrogram data of minmax normalized constructed data

getdendrodata AnalysisClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 
}}

