% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{getinitializedsubclassobjects}
\alias{getinitializedsubclassobjects}
\title{getinitializedsubclassobjects}
\usage{
getinitializedsubclassobjects(dataobject, parameterobject)
}
\arguments{
\item{dataobject}{A DataClass object that is used by computeValue method for the sub class object.}

\item{parameterobject}{A ParameterClass object that is composed of sub class names}
}
\description{
getinitializedsubclassobjects is a constructor function for constructing sub class objects
}

