% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getlofscores}
\alias{getlofscores}
\alias{getlofscores,AnalysisClass-method}
\title{getlofscores}
\usage{
getlofscores(object)

\S4method{getlofscores}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(numeric) vector of LOF scores
}
\description{
get LOF scores minmax normalized constructed data

getlofscores AnalysisClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 
}}

