% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getlofsumdata}
\alias{getlofsumdata}
\alias{getlofsumdata,AnalysisClass-method}
\title{getlofsumdata}
\usage{
getlofsumdata(object)

\S4method{getlofsumdata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) variable importance scores and feature names
}
\description{
get LOF scores of min-max normalized constructed data plus numerical imputed base data

getlofsumdata AnalysisClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 
}}

