% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getlongformatconstructeddata}
\alias{getlongformatconstructeddata}
\alias{getlongformatconstructeddata,AnalysisClass-method}
\title{get constructed data in long format}
\usage{
getlongformatconstructeddata(object)

\S4method{getlongformatconstructeddata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) long format constructed data
}
\description{
get constructed data in long format
}
\keyword{internal}

