% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getlongformatminmaxconstructeddata}
\alias{getlongformatminmaxconstructeddata}
\alias{getlongformatminmaxconstructeddata,AnalysisClass-method}
\title{getlongformatminmaxconstructeddata}
\usage{
getlongformatminmaxconstructeddata(object)

\S4method{getlongformatminmaxconstructeddata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) long format data
}
\description{
get long format of minmax normalized constructed data
}
\keyword{internal}

