% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getminmaxconstructeddata}
\alias{getminmaxconstructeddata}
\alias{getminmaxconstructeddata,AnalysisClass-method}
\alias{getminmaxconstructeddata,RunClass-method}
\title{getminmaxconstructeddata}
\usage{
getminmaxconstructeddata(object)

\S4method{getminmaxconstructeddata}{AnalysisClass}(object)

\S4method{getminmaxconstructeddata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) min-max normalized data
}
\description{
get contructed data that have been min-max normalized to [0,1]

getminmaxconstructeddata AnalysisClass

getminmaxconstructeddata RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 

\item \code{RunClass}: 
}}

