% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getname}
\alias{getname}
\alias{getname,AnalysisClass-method}
\alias{getname,BaseClass-method}
\title{getname}
\usage{
getname(object)

\S4method{getname}{AnalysisClass}(object)

\S4method{getname}{BaseClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or BaseClass)}
}
\value{
(character) name of the object
}
\description{
get name of an object

getname AnalysisClass

getname BaseClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 

\item \code{BaseClass}: 
}}

