% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getnumericombineddata}
\alias{getnumericombineddata}
\alias{getnumericombineddata,AnalysisClass-method}
\alias{getnumericombineddata,RunClass-method}
\title{getnumericombineddata}
\usage{
getnumericombineddata(object)

\S4method{getnumericombineddata}{AnalysisClass}(object)

\S4method{getnumericombineddata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) numeric columns combined data
}
\description{
get numeric columns of combined data

getnumericombineddata AnalysisClass

getnumericombineddata RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 

\item \code{RunClass}: 
}}

