% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{methods}
\name{getparameters}
\alias{getparameters}
\alias{getparameters,ParameterClass-method}
\alias{getparameters,SetUpClass-method}
\title{getparameters}
\usage{
getparameters(object)

\S4method{getparameters}{ParameterClass}(object)

\S4method{getparameters}{SetUpClass}(object)
}
\arguments{
\item{object}{(ParameterClass or SetUpClass)}
}
\description{
getparameters is a method for getting the list of parameters from a ParameterClass object or a Parameter object from SetUpClass object
}
\keyword{internal}

