% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getvariableimportancedata}
\alias{getvariableimportancedata}
\alias{getvariableimportancedata,AnalysisClass-method}
\title{get random forest variable importance data}
\usage{
getvariableimportancedata(object)

\S4method{getvariableimportancedata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) variable importance scores and feature names
}
\description{
get random forest variable importance data
}
\keyword{internal}

