% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\name{initializeanalysisclassobject}
\alias{initializeanalysisclassobject}
\title{initializeanalysisclassobject}
\usage{
initializeanalysisclassobject(object, dataobject)
}
\arguments{
\item{object}{A name of an object that contains the constructed feature vectors (i.e. initilized sub class objects)}

\item{dataobject}{A DataClass object that is used to computeValue in AnalysisClass computations}
}
\value{
An analysis class object
}
\description{
initializeanalysisclassobject is a constructor function for constructing AnalysisClass objects
}

