% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializecontrolclassobject}
\alias{initializecontrolclassobject}
\title{initializecontrolclassobject}
\usage{
initializecontrolclassobject(setups)
}
\arguments{
\item{setups}{(list) Name of SetUpClass objects}
}
\value{
(ControlClass) object
}
\description{
initializecontrolclassobject is a constructor function for intializing ControlClass objects.
}

