% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializedataobject}
\alias{initializedataobject}
\title{constructor function for initializing a DataClass object}
\usage{
initializedataobject(data)
}
\arguments{
\item{data}{(data frame)}
}
\value{
(DataClass)
}
\description{
constructor function for initializing a DataClass object
}

