% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializeparameterclassobject}
\alias{initializeparameterclassobject}
\title{constructor function for intializing a ParameterClass objects}
\usage{
initializeparameterclassobject(parameters)
}
\arguments{
\item{parameters}{(list) Name of sub classes}
}
\value{
(ParameterClass) object
}
\description{
constructor function for intializing a ParameterClass objects
}

