% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\name{initializeReportClass}
\alias{initializeReportClass}
\title{initializeReportClass is a constructor function for initializing a ReportClass object.}
\usage{
initializeReportClass(object)
}
\arguments{
\item{object}{(AnaysisClass)}
}
\description{
initializeReportClass is a constructor function for initializing a ReportClass object.
}

