% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializesetupclassobject}
\alias{initializesetupclassobject}
\title{initializesetupclassobject}
\usage{
initializesetupclassobject(objectname, parameterobject, dataobject)
}
\arguments{
\item{objectname}{(character) Name of the setup}

\item{parameterobject}{(ParameterClass)}

\item{dataobject}{(DataClass)}
}
\description{
initializesetupclassobject is a constructor function for initializing a SetUpClass object
}

