% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{initializesubclassobject}
\alias{initializesubclassobject}
\title{initializesubclassobject}
\usage{
initializesubclassobject(classname, dataobject)
}
\arguments{
\item{classname}{(character) A name of class of an object is initialized from (e.g. "MissingValueShare").}

\item{dataobject}{A DataClass object that is used by computeValue method for the sub class object.}
}
\description{
initializesubclassobject is a constructor function for constructing objects for classes inhereted from BaseClass
}

