% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\name{plotBAR}
\alias{plotBAR}
\title{plotBAR}
\usage{
plotBAR(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotBAR is a generic function for for plotting barplots of constructed features
}

