% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotCMDS}
\alias{plotBAR,ReportClass-method}
\alias{plotBAR,RunClass-method}
\alias{plotCMDS}
\alias{plotCMDS,ReportClass-method}
\alias{plotCMDS,RunClass-method}
\title{plotCMDS}
\usage{
plotCMDS(object)

\S4method{plotCMDS}{ReportClass}(object)

\S4method{plotCMDS}{RunClass}(object)

\S4method{plotBAR}{ReportClass}(object)

\S4method{plotBAR}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotCMDS is a generic function for plotting classical multidimensional scaling of constructed features

plotCMDS ReportClass

plotCMDS RunClass

plotBAR ReportClass

plotBAR RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 

\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

