% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotDENSITY}
\alias{plotDENSITY}
\alias{plotDENSITY,ReportClass-method}
\alias{plotDENSITY,RunClass-method}
\title{generic function for  plotting density estimates of constructed features}
\usage{
plotDENSITY(object)

\S4method{plotDENSITY}{ReportClass}(object)

\S4method{plotDENSITY}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for  plotting density estimates of constructed features
}

