% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotHEATMAP}
\alias{plotHEATMAP}
\alias{plotHEATMAP,ReportClass-method}
\alias{plotHEATMAP,RunClass-method}
\title{plotHEATMAP}
\usage{
plotHEATMAP(object)

\S4method{plotHEATMAP}{ReportClass}(object)

\S4method{plotHEATMAP}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotHEATMAP is a generic function for plotting heatmap of constructed features

plotHEATMAP ReportClass

plotHEATMAP RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

