% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotLOFSUM}
\alias{plotLOFSUM}
\alias{plotLOFSUM,ReportClass-method}
\alias{plotLOFSUM,RunClass-method}
\title{generic function for plotting lof sum of constructed features}
\usage{
plotLOFSUM(object)

\S4method{plotLOFSUM}{ReportClass}(object)

\S4method{plotLOFSUM}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting lof sum of constructed features
}

