% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotOUTLIERS}
\alias{plotOUTLIERS}
\alias{plotOUTLIERS,ReportClass-method}
\alias{plotOUTLIERS,RunClass-method}
\title{generic function for plotting density of LOF scores}
\usage{
plotOUTLIERS(object)

\S4method{plotOUTLIERS}{ReportClass}(object)

\S4method{plotOUTLIERS}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting density of LOF scores
}

