% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotSCATTERMATRIX}
\alias{plotSCATTERMATRIX}
\alias{plotSCATTERMATRIX,ReportClass-method}
\alias{plotSCATTERMATRIX,RunClass-method}
\title{plotSCATTERMATRIX}
\usage{
plotSCATTERMATRIX(object)

\S4method{plotSCATTERMATRIX}{ReportClass}(object)

\S4method{plotSCATTERMATRIX}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotSCATTERMATRIX is a generic function for plotting scatter matrix of constructed features.

plotSCATTERMATRIX ReportClass

plotSCATTERMATRIX RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

