% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotVARCLUST}
\alias{plotVARCLUST}
\alias{plotVARCLUST,ReportClass-method}
\alias{plotVARCLUST,RunClass-method}
\title{plotVARCLUST}
\usage{
plotVARCLUST(object)

\S4method{plotVARCLUST}{ReportClass}(object)

\S4method{plotVARCLUST}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotVARCLUST is a generic function for plotting variable clusters of constructed features

plotVARCLUST ReportClass

plotVARCLUST RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

