% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotVARIMP}
\alias{plotVARIMP}
\alias{plotVARIMP,ReportClass-method}
\alias{plotVARIMP,RunClass-method}
\title{plotVARIMP}
\usage{
plotVARIMP(object)

\S4method{plotVARIMP}{ReportClass}(object)

\S4method{plotVARIMP}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotHEATMAP is a generic function for variable importance (predicting the class labels in original data) of constructed features

plotVARIMP ReportClass

plotVARIMP RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

