% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06RunClass.R
\name{preproviz}
\alias{preproviz}
\title{preproviz}
\usage{
preproviz(controlobject)
}
\arguments{
\item{controlobject}{(data frame or ControlClass)}
}
\value{
(RunClass) object
}
\description{
preproviz is an execution function to run either a ControlClass object (complex setup)
or data frame (simple setup). Output can be plotted with plot-methods, e.g. plotVARCLUST(x)
}

