% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{class}
\name{ReportClass-class}
\alias{ReportClass-class}
\title{An S4 class representing visualizations}
\description{
An S4 class representing visualizations
}
\section{Slots}{

\describe{
\item{\code{density}}{density plot of all constructed features}

\item{\code{heatmap}}{heatmap}

\item{\code{cmds}}{classical multidimensional scaling}

\item{\code{variableclusters}}{hiearchical variable clustering}

\item{\code{outliers}}{LOF outlier scores}

\item{\code{varimp}}{variable importance}

\item{\code{lofsum}}{sum of LOF scores}
}}

