% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{RunClass-class}
\alias{RunClass-class}
\title{RunClass}
\value{
A RunClass object
}
\description{
RunClass is an class contain ReportClass and AnalysisClass objects as separate lists.
A RunClass object is the output of running the main function preproviz() and can be
studied with either get methods (AnalysisClass objects) or plot methods (ReportClass
objects).
}
\section{Slots}{

\describe{
\item{\code{reports}}{A list of ReportClass objects}

\item{\code{analysis}}{A list of AnalysisClass object}
}}

