% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01BaseClass.R
\docType{class}
\name{BaseClass-class}
\alias{BaseClass-class}
\title{An abstract S4 class representing contructed features.}
\description{
An abstract S4 class representing contructed features.
}
\section{Slots}{

\describe{
\item{\code{objectname}}{(character) name of the object}

\item{\code{valuevector}}{(numeric) constructed feature vector}

\item{\code{isvalid}}{(logical) result of object validation}

\item{\code{preimpute}}{(logical) whether valuevector iss computed before missing value imputation}
}}

