% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{computeValue}
\alias{computeValue}
\title{generic function for computing constructed feature vectors}
\usage{
computeValue(object, dataobject)
}
\arguments{
\item{object}{(sub class object inherited from BaseClass)}

\item{dataobject}{(DataClass)}
}
\value{
(numeric) feature vector
}
\description{
generic function for computing constructed feature vectors
}

