% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01BaseClass.R
\name{constructfeature}
\alias{constructfeature}
\title{constructfeature}
\usage{
constructfeature(classname, operation, mode = "all", impute = FALSE)
}
\arguments{
\item{classname}{(character) Name of the inherited class}

\item{operation}{(expression) Feature construction operation . The expression is evaluated by computevalue method.}

\item{mode}{(character) Mode of data to be used in construction . Defaults to "all", option "numeric" for numeric data without class labels.}

\item{impute}{(logical) Impute whether construction is done before missing value imputation . Defaults to "FALSE"}
}
\value{
(BaseClass) A sub class inhereted from BaseClass

A computevalue method for the sub class
}
\description{
constructfeature is a convenience function for inheriting a user-defined sub class from BaseClass and defining computevalue methods for them.
}

