% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{ControlClass-class}
\alias{ControlClass-class}
\title{A ControlClass is an class containing SetUpClass objects.}
\description{
A ControlClass is an class containing SetUpClass objects.
}
\section{Slots}{

\describe{
\item{\code{setups}}{A list of SetUpClass objects}
}}

