% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultControl.R
\docType{data}
\name{defaultParameters}
\alias{defaultParameters}
\title{defaultParameters}
\format{An object of class \code{ParameterClass} of length 1.}
\usage{
defaultParameters
}
\description{
defaultParameters include nine experimental constructed features (techically, subclasses)
}
\keyword{datasets}

