% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getcmdsdata}
\alias{getcmdsdata}
\alias{getcmdsdata,AnalysisClass-method}
\title{getcmdsdata}
\usage{
getcmdsdata(object)

\S4method{getcmdsdata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) cmds points
}
\description{
get classical multidimensional scaling two-dimensional data points from minmaxconstructed data

getcmdsdata AnalysisClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 
}}

