% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R, R/04ControlClass.R
\docType{methods}
\name{getcombineddata}
\alias{getcombineddata}
\alias{getcombineddata,AnalysisClass-method}
\alias{getcombineddata,RunClass-method}
\title{getcombineddata}
\usage{
getcombineddata(object)

\S4method{getcombineddata}{AnalysisClass}(object)

\S4method{getcombineddata}{RunClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or RunClass)}
}
\value{
(data frame) combined data
}
\description{
get basedata and constructed data combined

getcombineddata AnalysisClass

getcombineddata RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 

\item \code{RunClass}: 
}}

