% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getname}
\alias{getname}
\alias{getname,AnalysisClass-method}
\alias{getname,BaseClass-method}
\title{get name of an object}
\usage{
getname(object)

\S4method{getname}{AnalysisClass}(object)

\S4method{getname}{BaseClass}(object)
}
\arguments{
\item{object}{(AnalysisClass or BaseClass)}
}
\value{
(character) name of the object
}
\description{
get name of an object
}
\keyword{internal}

