% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03AnalysisClass.R
\docType{methods}
\name{getnumericbasedata}
\alias{getnumericbasedata}
\alias{getnumericbasedata,AnalysisClass-method}
\title{getnumericbasedata}
\usage{
getnumericbasedata(object)

\S4method{getnumericbasedata}{AnalysisClass}(object)
}
\arguments{
\item{object}{(AnalysisClass)}
}
\value{
(data frame) numeric columns of basedata
}
\description{
get numeric data of an object

getbasedata AnalysisClass
}
\section{Methods (by class)}{
\itemize{
\item \code{AnalysisClass}: 
}}

