% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializecontrolclassobject}
\alias{initializecontrolclassobject}
\title{constructor function for intializing a ControlClass object}
\usage{
initializecontrolclassobject(setups)
}
\arguments{
\item{setups}{(list) Name of SetUpClass objects}
}
\value{
(ControlClass) object
}
\description{
constructor function for intializing a ControlClass object
}

