% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializedataobject}
\alias{initializedataobject}
\title{initializedataObject}
\usage{
initializedataobject(data)
}
\arguments{
\item{data}{(data frame)}
}
\value{
(DataClass)
}
\description{
initializedataobject is a constructor function for initializing aDataClass object from original data frame
}

