% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializeparameterclassobject}
\alias{initializeparameterclassobject}
\title{initializeparameterclassobject}
\usage{
initializeparameterclassobject(parameters)
}
\arguments{
\item{parameters}{(list) Name of sub classes}
}
\value{
(ParameterClass) object
}
\description{
initializeparameterclassobject is a constructor function for intializing ParameterClass objects.
}

