% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\name{initializesetupclassobject}
\alias{initializesetupclassobject}
\title{constructor function for initializing a SetUpClass object}
\usage{
initializesetupclassobject(objectname, parameterobject, dataobject)
}
\arguments{
\item{objectname}{(character) Name of the setup}

\item{parameterobject}{(ParameterClass)}

\item{dataobject}{(DataClass)}
}
\description{
constructor function for initializing a SetUpClass object
}

