% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{ParameterClass-class}
\alias{ParameterClass-class}
\title{ParameterClass}
\description{
ParameterClass is a class containing a list of sub class objects (i.e. constructed features, inhereted
from BaseClass). A ParameterClass object in a SetUpClass object defines which constructed features are computed
from a DataClass object
}
\section{Slots}{

\describe{
\item{\code{parameters}}{A list of sub class objects}
}}

