% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotCMDS}
\alias{plotCMDS}
\alias{plotCMDS,ReportClass-method}
\alias{plotCMDS,RunClass-method}
\title{generic function for plotting classical multidimensional scaling}
\usage{
plotCMDS(object)

\S4method{plotCMDS}{ReportClass}(object)

\S4method{plotCMDS}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting classical multidimensional scaling
}

