% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotHEATMAP}
\alias{plotHEATMAP}
\alias{plotHEATMAP,ReportClass-method}
\alias{plotHEATMAP,RunClass-method}
\title{generic function for plotting heatmap}
\usage{
plotHEATMAP(object)

\S4method{plotHEATMAP}{ReportClass}(object)

\S4method{plotHEATMAP}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting heatmap
}

