% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotLOFSUM}
\alias{plotLOFSUM}
\alias{plotLOFSUM,ReportClass-method}
\alias{plotLOFSUM,RunClass-method}
\title{plotLOFSUM}
\usage{
plotLOFSUM(object)

\S4method{plotLOFSUM}{ReportClass}(object)

\S4method{plotLOFSUM}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotLOFSUM is a generic function for lof sum of constructed features

plotLOFSUM ReportClass

plotLOFSUM RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

