% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotOUTLIERS}
\alias{plotOUTLIERS}
\alias{plotOUTLIERS,ReportClass-method}
\alias{plotOUTLIERS,RunClass-method}
\title{plotOUTLIERS}
\usage{
plotOUTLIERS(object)

\S4method{plotOUTLIERS}{ReportClass}(object)

\S4method{plotOUTLIERS}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
plotOUTLIERS is a generic function for plotting density of LOF scores of constructed features

plotOUTLIERS ReportClass

plotOUTLIERS RunClass
}
\section{Methods (by class)}{
\itemize{
\item \code{ReportClass}: 

\item \code{RunClass}: 
}}

