% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotVARCLUST}
\alias{plotVARCLUST}
\alias{plotVARCLUST,ReportClass-method}
\alias{plotVARCLUST,RunClass-method}
\title{generic function for plotting variable clusters}
\usage{
plotVARCLUST(object)

\S4method{plotVARCLUST}{ReportClass}(object)

\S4method{plotVARCLUST}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting variable clusters
}

