% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{methods}
\name{plotVARIMP}
\alias{plotVARIMP}
\alias{plotVARIMP,ReportClass-method}
\alias{plotVARIMP,RunClass-method}
\title{generic function for plotting variable importance}
\usage{
plotVARIMP(object)

\S4method{plotVARIMP}{ReportClass}(object)

\S4method{plotVARIMP}{RunClass}(object)
}
\arguments{
\item{object}{(ReportClass or RunClass)}
}
\description{
generic function for plotting variable importance
}

