% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06RunClass.R
\name{preproviz}
\alias{preproviz}
\title{the MAIN execution function}
\usage{
preproviz(controlobject)
}
\arguments{
\item{controlobject}{(data frame/list/ControlClass object)}
}
\value{
(RunClass) object
}
\description{
for simple exploration preproviz() takes data frame (one factor variable, other variables numeric)
as an argument. Two data sets can be compared by providing them as a list. For complex setups a
ControlClass object can be passed as an argument. See Vignette for examples. The output can be
visualized with PLOT functions.
}
\examples{
## result1 <- preproviz(iris)
## plotDENSITY(result1)
##
## iris2 <- iris
## iris2[sample(1:150,30),1] <- NA
## result2 <- preproviz(list(iris, iris2))
## plotVARCLUST(result2)
}

