% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05ReportingClass.R
\docType{class}
\name{ReportClass-class}
\alias{ReportClass-class}
\title{A ReportClass is an class containing ggplot2 objects created based on an AnalysisClass object.}
\description{
A ReportClass is an class containing ggplot2 objects created based on an AnalysisClass object.
}
\section{Slots}{

\describe{
\item{\code{densityplot}}{}

\item{\code{smat}}{}

\item{\code{heatmap}}{}

\item{\code{multidimensionalscaling}}{}
}}

