% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{RunClass-class}
\alias{RunClass-class}
\title{An S4 class representing preproviz output (data and visualizations)}
\value{
A RunClass object
}
\description{
RunClass is an class contain ReportClass and AnalysisClass objects as separate lists.
A RunClass object is the output of running the main function preproviz() and can be
studied with either get or plot methods.
}
\section{Slots}{

\describe{
\item{\code{reports}}{A list of ReportClass objects}

\item{\code{analysis}}{A list of AnalysisClass object}
}}

