% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04ControlClass.R
\docType{class}
\name{SetUpClass-class}
\alias{SetUpClass-class}
\title{SetUpClass}
\description{
SetUpClass is an class containing a DataClass object, a ParameterClass object and the name of the SetUpClass object
}
\section{Slots}{

\describe{
\item{\code{data}}{(DataClass)}

\item{\code{parameters}}{(ParameterClass)}

\item{\code{objectname}}{(character)}
}}

